068-AzureOpenAIApps/Student/Resources/Challenge-00/ContosoAIAppsBackend/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions == 1.18.0
azure-identity == 1.16.0
azure-cosmos == 4.6.0
azure-servicebus == 7.12.1
azure-search-documents == 11.4.0
azure-ai-documentintelligence == 1.0.0b2
redis == 5.0.3
openai == 1.17.0
langchain == 0.1.16
langchain-community == 0.0.32
langchain-openai == 0.1.3